feat!: delete the ghostty pipeline, the wails runtime and the spike
Phase 5, frontend cut. The largest delete in the repo's history:

- The ghostty-web terminal and its entire private-patching layer are
  gone: render-pause, block-glyphs, font-metrics, glyph-atlas,
  getline-pool, scrollback-perf, row-paint, hidden-canvas, term-keys,
  term-links, term-dom, term-fit and their tests — all workarounds for
  ghostty-web 0.4.0 or the WebKitGTK webview, neither of which exists
  anymore. XtermTerminalView is renamed to TerminalView, absorbing the
  shared helpers (font preload, base64 decode, colors).
- @wailsio/runtime and the generated bindings are gone: service shapes
  live in lib/api-types.ts (hand-owned mirrors of the Go structs' JSON
  tags), the RPC endpoint comes exclusively from the page URL
  (?token=..., &backend=<port> under the Vite dev server), and
  app-events listens on the /events socket alone.
- terminal-choice flag deleted — there is one terminal now. The
  measurement spike (cmd/spike, spike.html, src/spike) served its
  purpose and is deleted as promised. term-perf trimmed to the
  transport/stall metrics the xterm path uses.
