--- frontend/spike.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>lich chromium spike</title>
-    <style>
-      html,
-      body {
-        height: 100%;
-        margin: 0;
-        background: #06070f;
-        overflow: hidden;
-      }
-      #term {
-        height: 100%;
-        width: 100%;
-      }
-      #stats {
-        position: fixed;
-        top: 4px;
-        right: 8px;
-        z-index: 9;
-        padding: 2px 8px;
-        border-radius: 4px;
-        background: rgba(0, 0, 0, 0.55);
-        color: #86efac;
-        font: 12px monospace;
-        pointer-events: none;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="stats">[spike] warming up…</div>
-    <div id="term"></div>
-    <script type="module" src="/src/spike/main.ts"></script>
-  </body>
-</html>
